requirements/local.txt
-r base.txt
django-extensions==3.1.3
ipython==8.14.0
django-debug-toolbar==4.1.0
transformers==4.30.2